+++ src/app/constants.js
@@ -1,2 +1,2 @@
-export const ASSET_VERSION = "2026-07-13-posters";
+export const ASSET_VERSION = "2026-07-13-upgrade";
 export const DEFAULT_TARGET_COUNT = 99;